Move crpyto code to a its own service
- Encrypt and Decrypt functions moved to crypto service
- Add secret_key to env.d.ts
