+++ db/config.ts
@@ -1,7 +1,5 @@
 import { defineDb, defineTable, column } from 'astro:db';
 import { createId } from '@paralleldrive/cuid2';
-import { createCipheriv, createDecipheriv, randomBytes } from 'node:crypto';
-import { Buffer } from 'buffer';
 
 // https://astro.build/db/config
 
@@ -46,15 +44,3 @@ export default defineDb({
     EmailLog,
   },
 });
-
-export const encryptBody = (body: string): string => {
-  const iv = randomBytes(16);
-  const cipher = createCipheriv(
-    'aes-256-cbc',
-    Buffer.from(process.env.SECRET_KEY || '____this_is_a_top_secret_key____'),
-    iv
-  );
-  let encrypted = cipher.update(body);
-  encrypted = Buffer.concat([encrypted, cipher.final()]);
-  return encrypted.toString();
-};
